Test fails on macOS runners
See https://github.com/upkie/vulp/issues/96 although the error here was
rather:

    tests/DictionaryTest.cpp:859:10: error: conversion from 'Dictionary' to 'size_t' (aka 'unsigned long') is ambiguous
      size_t before = serialized("logger")("last_size");
             ^        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    bazel-out/darwin-dbg/bin/include/palimpsest/_virtual_includes/dictionary/palimpsest/Dictionary.h:530:3: note: candidate function
      operator bool &() { return this->as<bool>(); }
    ...

+++ tests/DictionaryTest.cpp
@@ -856,7 +856,9 @@ TEST(Dictionary, SerializeSizeT) {
   size_t buffer_size = serialized.serialize(buffer);
   deserialized.update(buffer.data(), buffer_size);
 
-  size_t before = serialized("logger")("last_size");
+  // On some systems, conversion from 'Dictionary' to 'size_t' (typically
+  // 'unsigned long') is ambiguous, so we go directly for unsigned
+  unsigned before = serialized("logger")("last_size");
   unsigned after = deserialized("logger")("last_size");
   ASSERT_EQ(before, after);
 }
